tests/submission/PhilipTaylor/canvas/toDataURL.jpeg.alpha.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: toDataURL.jpeg.alpha</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.toDataURL.html">toDataURL</a>.<a href="index.toDataURL.jpeg.html">jpeg</a>.alpha</h1>
<p class="desc">toDataURL with JPEG composites onto black</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.toDataURL.jpeg">toDataURL.jpeg</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.toDataURL.noalpha">toDataURL.noalpha</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="toDataURL.jpeg.alpha.png">
</p><ul id="d"></ul>
<script>
var t = async_test("toDataURL with JPEG composites onto black");
_addTest(function(canvas, ctx) {

ctx.fillStyle = 'rgba(128, 255, 128, 0.5)';
ctx.fillRect(0, 0, 100, 50);
ctx.globalCompositeOperation = 'destination-over'; // should be ignored by toDataURL
var data = canvas.toDataURL('image/jpeg');
ctx.globalCompositeOperation = 'source-over';
if (!data.match(/^data:image\/jpeg[;,]/)) {
  _assert(true, "true");
} else {
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  var img = new Image();
  deferTest();
  img.onload = wrapFunction(function ()
  {
      ctx.drawImage(img, 0, 0);
      
      _assertPixelApprox(canvas, 50,25, 63,127,63,255, "50,25", "63,127,63,255", 8);
  });
  img.src = data;
}


});
</script>

<div id="log"></div>
</body></html>
